darknet frontend changes, review comments reworked.
