+++ python/nnvm/testing/__init__.py
@@ -7,3 +7,4 @@ from . import mobilenet
 from . import mlp
 from . import resnet
 from . import vgg
+from . import darknet
