+++ python/nnvm/top/__init__.py
@@ -7,6 +7,7 @@ from . import tensor
 from . import nn
 from . import transform
 from . import reduction
+from . import vision
 
 from .registry import OpPattern
 from .registry import register_compute, register_schedule, register_pattern
