+++ python/nnvm/top/nn.py
@@ -194,49 +194,3 @@ def schedule_upsampling(_, outs, target):
         return topi.generic.schedule_injective(outs)
 
 reg.register_pattern("upsampling", OpPattern.INJECTIVE)
-
-@reg.register_compute("shortcut")
-def compute_shortcut(attrs, inputs, _):
-    """Compute definition of shortcut"""
-    out = topi.nn.vision.shortcut(inputs[0], inputs[1])
-    return out
-
-@reg.register_schedule("shortcut")
-def schedule_shortcut(attrs, outs, target):
-    """Schedule definition of shortcut"""
-    with tvm.target.create(target):
-        return topi.generic.schedule_shortcut(outs)
-
-reg.register_pattern("shortcut", OpPattern.OUT_ELEMWISE_FUSABLE)
-
-@reg.register_compute("reorg")
-def compute_reorg(attrs, inputs, _):
-    """Compute definition of reorg"""
-    return topi.nn.vision.reorg(inputs[0], attrs.get_int("stride"))
-
-@reg.register_schedule("reorg")
-def schedule_reorg(attrs, outs, target):
-    """Schedule definition of reorg"""
-    with tvm.target.create(target):
-        return topi.generic.schedule_reorg(outs)
-
-reg.register_pattern("reorg", OpPattern.OUT_ELEMWISE_FUSABLE)
-
-@reg.register_compute("region")
-def compute_region(attrs, inputs, _):
-    """Compute definition of region"""
-    n = attrs.get_int("n")
-    classes = attrs.get_int("classes")
-    coords = attrs.get_int("coords")
-    background = attrs.get_int("background")
-    softmax = attrs.get_int("softmax")
-
-    return topi.nn.vision.region(inputs[0], n, classes, coords, background, softmax)
-
-@reg.register_schedule("region")
-def schedule_region(attrs, outs, target):
-    """Schedule definition of region"""
-    with tvm.target.create(target):
-        return topi.generic.schedule_region(outs)
-
-reg.register_pattern("region", OpPattern.OUT_ELEMWISE_FUSABLE)
